Drop loftctl/v4/pkg/config dep
