+++ pkg/util/clihelper/clihelper.go
@@ -13,7 +13,6 @@ import (
 	"github.com/loft-sh/log"
 	"github.com/loft-sh/vcluster/pkg/cli/find"
 	"github.com/loft-sh/vcluster/pkg/cli/podprinter"
-	"github.com/loft-sh/vcluster/pkg/platform/kubeconfig"
 	utilkubeconfig "github.com/loft-sh/vcluster/pkg/util/kubeconfig"
 	"github.com/pkg/errors"
 	"golang.org/x/mod/semver"
@@ -41,40 +40,6 @@ var CriticalStatus = map[string]bool{
 
 var SortPodsByNewest = func(pods []corev1.Pod, i, j int) bool {
 	return pods[i].CreationTimestamp.Unix() > pods[j].CreationTimestamp.Unix()
-}
-
-// GetProKubeConfig builds a pro kube config from options and client
-func GetProKubeConfig(options kubeconfig.ContextOptions) (*clientcmdapi.Config, error) {
-	contextName := options.Name
-	cluster := clientcmdapi.NewCluster()
-	cluster.Server = options.Server
-	cluster.CertificateAuthorityData = options.CaData
-	cluster.InsecureSkipTLSVerify = options.InsecureSkipTLSVerify
-
-	authInfo := clientcmdapi.NewAuthInfo()
-	if options.Token != "" || options.ClientCertificateData != nil || options.ClientKeyData != nil {
-		authInfo.Token = options.Token
-		authInfo.ClientKeyData = options.ClientKeyData
-		authInfo.ClientCertificateData = options.ClientCertificateData
-	}
-
-	config := clientcmdapi.NewConfig()
-	config.Clusters[contextName] = cluster
-	config.AuthInfos[contextName] = authInfo
-
-	// Update kube context
-	kubeContext := clientcmdapi.NewContext()
-	kubeContext.Cluster = contextName
-	kubeContext.AuthInfo = contextName
-	kubeContext.Namespace = options.CurrentNamespace
-
-	config.Contexts[contextName] = kubeContext
-	config.CurrentContext = contextName
-
-	// set kind & version
-	config.APIVersion = "v1"
-	config.Kind = "Config"
-	return config, nil
 }
 
 // GetKubeConfig attempts to read the kubeconfig from the default Secret and
